+++ tests/elf_binary_printer_test.py
@@ -553,10 +553,13 @@ class ElfBinaryPrinterTests(BinaryPPrinterTest):
             with self.subTest(subtest=subtest):
                 self.subtest_dynamic_entries(*subtest)
 
-    def test_export_dynamic(self):
+    def subtest_export_dynamic(self, global_not_in_dynsym: bool):
         """
         Test that we pass -Wl,--export-dynamic when all
         global visible symbols are in .dynsym.
+        Also, when not all global visible symbols are in .dynsym,
+        test that we generate a dynamic-list file, and pass
+        -Wl,--dynamic-list with it.
         """
 
         ir, module = gth.create_test_module(
@@ -619,10 +622,34 @@ class ElfBinaryPrinterTests(BinaryPPrinterTest):
                 (".symtab", index)
             ]
 
+        if global_not_in_dynsym:
+            block = gth.add_code_block(section_bi, code_bytes, {})
+
+            symbol = gth.add_symbol(module, "f6_symbol", block)
+            module.aux_data["elfSymbolInfo"].data[symbol.uuid] = (
+                0,
+                "FUNC",
+                "GLOBAL",
+                "DEFAULT",
+                0,
+            )
+            module.aux_data["elfSymbolTabIdxInfo"].data[symbol.uuid] = [
+                (".symtab", 5)
+            ]
+
         module.aux_data["libraries"].data.extend(["libc.so"])
 
         # Build binary
         with self.binary_print(ir) as result:
+            print(result.completed_process.stdout)
+            if global_not_in_dynsym:
+                self.assertIn(
+                    "-Wl,--dynamic-list=", result.completed_process.stdout
+                )
+            else:
+                self.assertIn(
+                    "-Wl,--export-dynamic", result.completed_process.stdout
+                )
             dynsym = self.readelf(result.path, "--dyn-syms")
             symtab = self.readelf(result.path, "--syms")
 
@@ -636,12 +663,27 @@ class ElfBinaryPrinterTests(BinaryPPrinterTest):
             # The hidden global are not exported
             self.assertNotIn("f4_symbol", dynsym.stdout)
             self.assertNotIn("f5_symbol", dynsym.stdout)
-            # The hidden global have been transformed to local
-            self.assert_readelf_syms(
-                symtab.stdout,
-                ("FUNC", "LOCAL", "DEFAULT", "f4_symbol"),
-                ("FUNC", "LOCAL", "DEFAULT", "f5_symbol"),
-            )
+            if not global_not_in_dynsym:
+                # The hidden global have been transformed to local
+                self.assert_readelf_syms(
+                    symtab.stdout,
+                    ("FUNC", "LOCAL", "DEFAULT", "f4_symbol"),
+                    ("FUNC", "LOCAL", "DEFAULT", "f5_symbol"),
+                )
+
+    def test_export_dynamic(self):
+        """
+        Set up subtests for export-dynamic and dynamic-list
+        """
+        subtests = (
+            # global_not_in_dynsym?
+            (True),
+            (False),
+        )
+
+        for global_not_in_dynsym in subtests:
+            with self.subTest(global_not_in_dynsym=global_not_in_dynsym):
+                self.subtest_export_dynamic(global_not_in_dynsym)
 
     def subtest_elf_stack_properties(self, stack_size: int, stack_exec: bool):
         """